Refactor home
